BackEnd/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
pydantic==2.11.3
pydantic-settings==2.8.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
python-dotenv==1.1.0
google-generativeai==0.8.5
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.26.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-api-python-client==2.167.0
google-cloud-aiplatform==1.89.0
google-genai==1.11.0
aiofiles==23.2.1
httpx==0.28.1
starlette==0.46.2
gunicorn==21.2.0
pytest==7.4.3
requests==2.32.3
websockets==15.0.1
tqdm==4.67.1
langchain==0.3.25
langchain-community==0.3.23
chromadb==1.0.8